+++ src/Router/tests/Stub/RoutesTestCore.php
@@ -13,6 +13,6 @@ namespace Spiral\Tests\Router\Stub;
 
 use Spiral\Core\AbstractCore;
 
-class AnnotatedRoutesTestCore extends AbstractCore
+class RoutesTestCore extends AbstractCore
 {
 }