+++ src/SendIt/src/Config/MailerConfig.php
@@ -21,7 +21,8 @@ final class MailerConfig extends InjectableConfig
     protected $config = [
         'dsn' => '',
         'from' => '',
-        'pipeline' => '',
+        'queue' => null,
+        'pipeline' => null,
         'queueConnection' => null,
     ];
 
@@ -35,9 +36,17 @@ final class MailerConfig extends InjectableConfig
         return $this->config['from'];
     }
 
-    public function getQueuePipeline(): string
+    /**
+     * @deprecated since v2.9.
+     */
+    public function getQueuePipeline(): ?string
     {
-        return $this->config['pipeline'];
+        return $this->getQueue();
+    }
+
+    public function getQueue(): ?string
+    {
+        return $this->config['queue'] ?? $this->config['pipeline'] ?? null;
     }
 
     public function getQueueConnection(): ?string