+++ CHANGELOG.md
@@ -11,6 +11,14 @@
 - **Other Features**
   - Class `Spiral\Mailer\Message` is no longer final and is available for extension
 
+## v2.9.1 - 2022-02-11
+
+- **High Impact Changes**
+- **Medium Impact Changes**
+  - [spiral/sendit] Method `getQueuePipeline` of `Spiral\SendIt\Config\MailerConfig` class is deprecated.
+    Use method `getQueue` instead. Added environment variables `MAILER_QUEUE` and `MAILER_QUEUE_CONNECTION`
+- **Other Features**
+- 
 ## v2.9.0 - 2022-02-03
 
 - **High Impact Changes**